structure new components under comman
