+++ src/components/common/Theme.tsx
@@ -1,0 +1,84 @@
+export const theme = {
+  space: {
+    px: '1px',
+    0: '0',
+    0.5: '0.125rem',
+    1: '0.25rem',
+    1.5: '0.375rem',
+    2: '0.5rem',
+    2.5: '0.625rem',
+    3: '0.75rem',
+    3.5: '0.875rem',
+    4: '1rem',
+    5: '1.25rem',
+    6: '1.5rem',
+    7: '1.75rem',
+    8: '2rem',
+    9: '2.25rem',
+    10: '2.5rem',
+    12: '3rem',
+    14: '3.5rem',
+    16: '4rem',
+    20: '5rem',
+    24: '6rem',
+    28: '7rem',
+    32: '8rem',
+    36: '9rem',
+    40: '10rem',
+    44: '11rem',
+    48: '12rem',
+    52: '13rem',
+    56: '14rem',
+    60: '15rem',
+    64: '16rem',
+    72: '18rem',
+    80: '20rem',
+    96: '24rem',
+  },
+  fonts: {
+    heading: 'Cooper Hewitt',
+    body: 'Roboto',
+    mono: 'Roboto Mono',
+  },
+  fontSizes: {
+    xs: '0.75rem',
+    sm: '0.875rem',
+    md: '1rem',
+    lg: '1.125rem',
+    xl: '1.25rem',
+    '2xl': '1.5rem',
+    '3xl': '1.875rem',
+    '4xl': '2.25rem',
+    '5xl': '3rem',
+    '6xl': '3.75rem',
+    '7xl': '4.5rem',
+    '8xl': '6rem',
+    '9xl': '8rem',
+  },
+  sizes: {
+    lg: {
+      h: 12,
+      minW: 12,
+      fontSize: 'lg',
+      px: 6,
+    },
+    md: {
+      h: 10,
+      minW: 10,
+      fontSize: 'md',
+      px: 4,
+    },
+    sm: {
+      h: 8,
+      minW: 8,
+      fontSize: 'sm',
+      px: 3,
+    },
+    xs: {
+      h: 6,
+      minW: 6,
+      fontSize: 'xs',
+      px: 2,
+    },
+  },
+};
