+++ src/components/common/index.tsx
@@ -1,0 +1,13 @@
+import Box from './Box';
+
+export { Box };
+export * from './Button';
+export * from './Fonts';
+export * from './Icon';
+export * from './Input';
+export * from './Link';
+export * from './Menu';
+export * from './Modal';
+export * from './Stack';
+export * from './Text';
+export * from './Theme';
